+++ src/pages/Documents/components/UploadForm.tsx
@@ -3,182 +3,48 @@
  * T-49-ST2: File upload with drag & drop, validation, progress
  */
 
-import React, { useState, useCallback, useRef } from 'react';
+import React from 'react';
+import { useFileUpload } from '../hooks/useFileUpload';
+import DropZone from './DropZone';
+import FilePreview from './FilePreview';
 
 interface UploadFormProps {
   onUploadSuccess?: (documentId: string) => void;
   onUploadError?: (error: string) => void;
 }
 
-const ALLOWED_FILE_TYPES = [
-  'application/pdf',
-  'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-  'text/markdown',
-];
-const ALLOWED_EXTENSIONS = ['.pdf', '.docx', '.md'];
-const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB
-
+// eslint-disable-next-line max-lines-per-function
 const UploadForm: React.FC<UploadFormProps> = ({ onUploadSuccess, onUploadError }) => {
-  const [isDragging, setIsDragging] = useState(false);
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-  const [isUploading, setIsUploading] = useState(false);
-  const [uploadProgress, setUploadProgress] = useState(0);
-  const [error, setError] = useState<string | null>(null);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  // Log props to avoid TypeScript unused warning (backend integration pending)
-  if (onUploadSuccess || onUploadError) {
-    console.debug('Upload callbacks ready for backend integration');
-  }
-
-  const validateFile = (file: File): string | null => {
-    if (
-      !ALLOWED_FILE_TYPES.includes(file.type) &&
-      !ALLOWED_EXTENSIONS.some(ext => file.name.endsWith(ext))
-    ) {
-      return `Invalid file type. Only PDF, DOCX, and MD files are allowed.`;
-    }
-    if (file.size > MAX_FILE_SIZE) {
-      return `File size exceeds ${MAX_FILE_SIZE / 1024 / 1024}MB limit.`;
-    }
-    return null;
-  };
-
-  const handleFileSelect = useCallback((file: File) => {
-    const validationError = validateFile(file);
-    if (validationError) {
-      setError(validationError);
-      setSelectedFile(null);
-      return;
-    }
-    setError(null);
-    setSelectedFile(file);
-  }, []);
-
-  const handleDragEnter = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  }, []);
-
-  const handleDragLeave = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  }, []);
-
-  const handleDragOver = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  }, []);
-
-  const handleDrop = useCallback(
-    (e: React.DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      setIsDragging(false);
-
-      const files = Array.from(e.dataTransfer.files);
-      if (files.length > 0) {
-        handleFileSelect(files[0]);
-      }
-    },
-    [handleFileSelect]
-  );
-
-  const handleFileInputChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const files = e.target.files;
-      if (files && files.length > 0) {
-        handleFileSelect(files[0]);
-      }
-    },
-    [handleFileSelect]
-  );
-
-  const handleUpload = async () => {
-    if (!selectedFile) return;
-
-    setIsUploading(true);
-    setUploadProgress(0);
-    setError(null);
-
-    // TODO: Implement actual upload to POST /api/upload endpoint
-    // For now, simulate upload progress
-    const interval = setInterval(() => {
-      setUploadProgress(prev => {
-        if (prev >= 100) {
-          clearInterval(interval);
-          return 100;
-        }
-        return prev + 10;
-      });
-    }, 200);
-
-    setTimeout(() => {
-      clearInterval(interval);
-      setIsUploading(false);
-      setUploadProgress(100);
-      setError('Upload endpoint not implemented yet. Backend POST /api/upload is pending.');
-      onUploadError?.('Upload endpoint not implemented');
-    }, 2000);
-  };
-
-  const handleBrowseClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleClearFile = () => {
-    setSelectedFile(null);
-    setError(null);
-    setUploadProgress(0);
-    if (fileInputRef.current) {
-      fileInputRef.current.value = '';
-    }
-  };
+  const {
+    isDragging,
+    selectedFile,
+    isUploading,
+    uploadProgress,
+    error,
+    fileInputRef,
+    handleDragEnter,
+    handleDragLeave,
+    handleDragOver,
+    handleDrop,
+    handleFileInputChange,
+    handleUpload,
+    handleBrowseClick,
+    handleClearFile,
+  } = useFileUpload({ onUploadSuccess, onUploadError });
 
   return (
     <div className="bg-white dark:bg-gray-800 rounded-lg shadow p-6">
       <h2 className="text-xl font-bold text-gray-900 dark:text-white mb-4">Upload Document</h2>
 
       {/* Drag & Drop Zone */}
-      <div
+      <DropZone
+        isDragging={isDragging}
         onDragEnter={handleDragEnter}
         onDragLeave={handleDragLeave}
         onDragOver={handleDragOver}
         onDrop={handleDrop}
-        className={`border-2 border-dashed rounded-lg p-8 text-center transition-colors ${
-          isDragging
-            ? 'border-blue-500 bg-blue-50 dark:bg-blue-900/20'
-            : 'border-gray-300 dark:border-gray-600'
-        }`}
-      >
-        <svg
-          className="mx-auto h-12 w-12 text-gray-400"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12"
-          />
-        </svg>
-        <p className="mt-2 text-sm text-gray-600 dark:text-gray-400">
-          Drag and drop your file here, or{' '}
-          <button
-            onClick={handleBrowseClick}
-            className="text-blue-600 dark:text-blue-400 hover:underline"
-          >
-            browse
-          </button>
-        </p>
-        <p className="mt-1 text-xs text-gray-500 dark:text-gray-500">
-          PDF, DOCX, MD files up to 10MB
-        </p>
-      </div>
+        onBrowseClick={handleBrowseClick}
+      />
 
       {/* Hidden File Input */}
       <input
@@ -191,59 +57,12 @@ const UploadForm: React.FC<UploadFormProps> = ({ onUploadSuccess, onUploadError 
 
       {/* Selected File Preview */}
       {selectedFile && (
-        <div className="mt-4 p-4 bg-gray-50 dark:bg-gray-700 rounded-lg">
-          <div className="flex items-center justify-between">
-            <div className="flex items-center space-x-3">
-              <svg
-                className="h-8 w-8 text-blue-600 dark:text-blue-400"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
-                />
-              </svg>
-              <div>
-                <p className="text-sm font-medium text-gray-900 dark:text-white">
-                  {selectedFile.name}
-                </p>
-                <p className="text-xs text-gray-500 dark:text-gray-400">
-                  {(selectedFile.size / 1024).toFixed(2)} KB
-                </p>
-              </div>
-            </div>
-            <button
-              onClick={handleClearFile}
-              className="text-gray-400 hover:text-gray-600 dark:hover:text-gray-300"
-            >
-              <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            </button>
-          </div>
-
-          {/* Progress Bar */}
-          {isUploading && (
-            <div className="mt-3">
-              <div className="w-full bg-gray-200 dark:bg-gray-600 rounded-full h-2">
-                <div
-                  className="bg-blue-600 h-2 rounded-full transition-all duration-300"
-                  style={{ width: `${uploadProgress}%` }}
-                />
-              </div>
-              <p className="text-xs text-gray-500 dark:text-gray-400 mt-1">{uploadProgress}%</p>
-            </div>
-          )}
-        </div>
+        <FilePreview
+          file={selectedFile}
+          isUploading={isUploading}
+          uploadProgress={uploadProgress}
+          onClear={handleClearFile}
+        />
       )}
 
       {/* Error Message */}
